Add support for long vectors (> 2^31-1)
R-3.0.0 added support for long vectors. The xts and zoo C code were
not updated to take advantage of this feature. This change in zoo is
necessary because xts C code supports long vectors (as-of 2023-03-06)
and calls the lag() and coredata() C functions in zoo.

See https://github.com/joshuaulrich/xts/issues/239

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/zoo@1185 edb9625f-4e0d-4859-8d74-9fd3b1da38cb